김영한 JPA 활용1 - 24. 주문 서비스 개발

+++ jpashop/src/main/java/nkm/study/jpabook/jpashop/repository/OrderRepository.java
@@ -16,7 +16,7 @@ public class OrderRepository {
         em.persist(order);
     }
 
-    public Order findOne(int orderId) {
+    public Order findOne(Long orderId) {
         return em.find(Order.class, orderId);
     }
